test: migrate zero-user-connectors-update to bdd
